Fix localstorage useeffect

+++ frontend/src/features/Favorite/Favorite.tsx
@@ -16,6 +16,7 @@ export const Favorite = ({ session }: Props) => {
   const { data: favoriteRecipes, isLoading } = useQuery({
     queryKey: ["favoriteRecipes"],
     queryFn: async () => {
+      if (!session?.access_token) return
       const response = await fetch(getUserFavoritesApi(), {
         headers: { Authorization: `Bearer ${session?.access_token}` },
       })
@@ -28,6 +29,7 @@ export const Favorite = ({ session }: Props) => {
   // NOTE: https://www.notion.so/utcode/JWT-4743f0e6a64e4ee7848818c9bc0efee1?pvs=4
   const onClickDeleteFavorite = useMutation({
     mutationFn: async (recipeId: number) => {
+      if (!session?.access_token) return
       const response = await fetch(deleteUserFavoritesApi(recipeId), {
         method: "DELETE",
         headers: { Authorization: `Bearer ${session?.access_token}` },
@@ -56,6 +58,7 @@ export const Favorite = ({ session }: Props) => {
             ))}
           </ul>
         ) : (
+          // TODO: サインインしてないときに表示を変える
           <p>お気に入りはまだありません。ハートボタンを押して追加してみましょう。</p>
         )}
         <Link to="/home">
